+++ eveuniverse/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "eveuniverse.apps.EveuniverseConfig"
 
-__version__ = "0.7.6"
+__version__ = "0.8.0a2"
 __title__ = "Eve Universe"