+++ eveuniverse/managers.py
@@ -1,8 +1,12 @@
 from collections import namedtuple
 import datetime as dt
 import logging
-from typing import List, Iterable, Tuple, Optional
+from typing import List, Iterable, Optional, Set, Tuple
+from urllib.parse import urljoin
 
+import requests
+
+from django.core.cache import cache
 from django.db import models, transaction
 from django.db.utils import IntegrityError
 from django.utils.timezone import now
@@ -20,12 +24,16 @@ logger = LoggerAddTag(logging.getLogger(__name__), __title__)
 
 FakeResponse = namedtuple("FakeResponse", ["status_code"])
 
+SDE_ZZEVE_URL = "https://sde.zzeve.com"
+
 
 class EveUniverseBaseModelManager(models.Manager):
-    def _defaults_from_esi_obj(self, eve_data_obj: dict) -> dict:
+    def _defaults_from_esi_obj(
+        self, eve_data_obj: dict, enabled_sections: Set[str] = None
+    ) -> dict:
         """compiles defaults from an esi data object for update/creating the model"""
         defaults = dict()
-        for field_name, mapping in self.model._esi_mapping().items():
+        for field_name, mapping in self.model._esi_mapping(enabled_sections).items():
             if not mapping.is_pk:
                 if not isinstance(mapping.esi_name, tuple):
                     if mapping.esi_name in eve_data_obj:
@@ -89,24 +97,31 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
             id: Eve Online ID of object
             include_children: if child objects should be updated/created as well (only when a new object is created)
             wait_for_children: when true child objects will be updated/created blocking (if any), else async (only when a new object is created)
-            enabled_sections: Sections to load regardless of current settings, e.g. `EveUniverseEntityModel.LOAD_DOGMAS` will always load dogmas for EveTypes (only when a new object is created)
+            enabled_sections: Sections to load regardless of current settings, e.g. `[EveType.Section.DOGMAS]` will always load dogmas for EveTypes
 
         Returns:
             A tuple consisting of the requested object and a created flag
         """
         id = int(id)
+        enabled_sections = self.model._enabled_sections_union(enabled_sections)
         try:
-            obj = self.get(id=id)
-            created = False
+            enabled_sections_filter = self._enabled_sections_filter(enabled_sections)
+            obj = self.filter(**enabled_sections_filter).get(id=id)
+            return obj, False
         except self.model.DoesNotExist:
-            obj, created = self.update_or_create_esi(
+            return self.update_or_create_esi(
                 id=id,
                 include_children=include_children,
                 wait_for_children=wait_for_children,
                 enabled_sections=enabled_sections,
             )
 
-        return obj, created
+    def _enabled_sections_filter(self, enabled_sections: Iterable[str]) -> dict:
+        return {
+            "enabled_sections": getattr(self.model.enabled_sections, section)
+            for section in enabled_sections
+            if str(section) in self.model.Section.values()
+        }
 
     def update_or_create_esi(
         self,
@@ -123,21 +138,29 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
             id: Eve Online ID of object
             include_children: if child objects should be updated/created as well (if any)
             wait_for_children: when true child objects will be updated/created blocking (if any), else async
-            enabled_sections: Sections to load regardless of current settings, e.g. `EveUniverseEntityModel.LOAD_DOGMAS` will always load dogmas for EveTypes
+            enabled_sections: Sections to load regardless of current settings, e.g. `[EveType.Section.DOGMAS]` will always load dogmas for EveTypes
 
         Returns:
             A tuple consisting of the requested object and a created flag
         """
         id = int(id)
-        enabled_sections = set(enabled_sections) if enabled_sections else set()
         add_prefix = make_logger_prefix("%s(id=%s)" % (self.model.__name__, id))
+        enabled_sections = self.model._enabled_sections_union(enabled_sections)
         try:
             eve_data_obj = self._transform_esi_response_for_list_endpoints(
-                id, self._fetch_from_esi(id)
+                id, self._fetch_from_esi(id=id, enabled_sections=enabled_sections)
             )
             if eve_data_obj:
-                defaults = self._defaults_from_esi_obj(eve_data_obj)
+                defaults = self._defaults_from_esi_obj(eve_data_obj, enabled_sections)
                 obj, created = self.update_or_create(id=id, defaults=defaults)
+                if enabled_sections and hasattr(obj, "enabled_sections"):
+                    updated_sections = False
+                    for section in enabled_sections:
+                        if str(section) in self.model.Section.values():
+                            setattr(obj.enabled_sections, section, True)
+                            updated_sections = True
+                    if updated_sections:
+                        obj.save()
                 inline_objects = self.model._inline_objects(enabled_sections)
                 if inline_objects:
                     self._update_or_create_inline_objects(
@@ -145,6 +168,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                         parent_obj=obj,
                         inline_objects=inline_objects,
                         wait_for_children=wait_for_children,
+                        enabled_sections=enabled_sections,
                     )
                 if include_children:
                     self._update_or_create_children(
@@ -168,7 +192,9 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
 
         return obj, created
 
-    def _fetch_from_esi(self, id: int = None) -> object:
+    def _fetch_from_esi(
+        self, id: int = None, enabled_sections: Iterable[str] = None
+    ) -> dict:
         """make request to ESI and return response data.
         Can handle raw ESI response from both list and normal endpoints.
         """
@@ -208,8 +234,9 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
         parent_obj: models.Model,
         inline_objects: dict,
         wait_for_children: bool,
+        enabled_sections: Iterable[str],
     ) -> None:
-        """updates_or_creates eve objects that are returns "inline" from ESI
+        """updates_or_creates eve objects that are returned "inline" from ESI
         for the parent eve objects as defined for this parent model (if any)
         """
         from .tasks import (
@@ -265,6 +292,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                             other_pk_info=other_pk_info,
                             parent2_model_name=parent2_model_name,
                             inline_model_name=model_name,
+                            enabled_sections=enabled_sections,
                         )
                     else:
                         task_update_or_create_inline_object(
@@ -275,6 +303,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                             parent2_model_name=parent2_model_name,
                             inline_model_name=model_name,
                             parent_model_name=type(parent_obj).__name__,
+                            enabled_sections=list(enabled_sections),
                         )
 
     def _update_or_create_inline_object(
@@ -285,6 +314,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
         other_pk_info: dict,
         parent2_model_name: str,
         inline_model_name: str,
+        enabled_sections: Iterable[str],
     ):
         """Updates or creates a single inline object.
         Will automatically create additional parent objects as needed
@@ -330,9 +360,11 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                 % self.model.__name__,
             )
 
-        for key, child_class in self.model._children().items():
+        for key, child_class in self.model._children(enabled_sections).items():
             if key in parent_eve_data_obj and parent_eve_data_obj[key]:
-                for id in parent_eve_data_obj[key]:
+                for obj in parent_eve_data_obj[key]:
+                    # TODO: Refactor this hack
+                    id = obj["planet_id"] if key == "planets" else obj
                     if wait_for_children:
                         ChildClass = self.model.get_model_class(child_class)
                         ChildClass.objects.update_or_create_esi(
@@ -356,6 +388,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
         *,
         include_children: bool = False,
         wait_for_children: bool = True,
+        enabled_sections: Iterable[str] = None,
     ) -> None:
         """updates or creates all objects of this class from ESI.
 
@@ -364,16 +397,20 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
         Args:
             include_children: if child objects should be updated/created as well (if any)
             wait_for_children: when false all objects will be loaded async, else blocking
+            enabled_sections: Sections to load regardless of current settings
         """
         from .tasks import update_or_create_eve_object
 
         add_prefix = make_logger_prefix(f"{self.model.__name__}")
+        enabled_sections = self.model._enabled_sections_union(enabled_sections)
         if self.model._is_list_only_endpoint():
             try:
                 esi_pk = self.model._esi_pk()
                 for eve_data_obj in self._fetch_from_esi():
                     args = {"id": eve_data_obj[esi_pk]}
-                    args["defaults"] = self._defaults_from_esi_obj(eve_data_obj)
+                    args["defaults"] = self._defaults_from_esi_obj(
+                        eve_data_obj=eve_data_obj, enabled_sections=enabled_sections
+                    )
                     self.update_or_create(**args)
 
             except Exception as ex:
@@ -395,6 +432,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                             id=id,
                             include_children=include_children,
                             wait_for_children=wait_for_children,
+                            enabled_sections=enabled_sections,
                         )
                     else:
                         update_or_create_eve_object.delay(
@@ -402,6 +440,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
                             entity_id=id,
                             include_children=include_children,
                             wait_for_children=wait_for_children,
+                            enabled_sections=list(enabled_sections),
                         )
             else:
                 raise TypeError(
@@ -414,6 +453,7 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
         ids: List[int],
         include_children: bool = False,
         wait_for_children: bool = True,
+        enabled_sections: Iterable[str] = None,
     ) -> models.QuerySet:
         """Gets or creates objects in bulk.
 
@@ -424,36 +464,44 @@ class EveUniverseEntityModelManager(EveUniverseBaseModelManager):
             ids: List of valid IDs of Eve objects
             include_children: when needed to updated/created if child objects should be updated/created as well (if any)
             wait_for_children: when true child objects will be updated/created blocking (if any), else async
+            enabled_sections: Sections to load regardless of current settings
 
         Returns:
             Queryset with all requested eve objects
         """
         ids = set(ids)
-        existing_ids = set(self.filter(id__in=ids).values_list("id", flat=True))
+        enabled_sections = self.model._enabled_sections_union(enabled_sections)
+        enabled_sections_filter = self._enabled_sections_filter(enabled_sections)
+        existing_ids = set(
+            self.filter(id__in=ids)
+            .filter(**enabled_sections_filter)
+            .values_list("id", flat=True)
+        )
         for id in ids.difference(existing_ids):
             self.update_or_create_esi(
                 id=int(id),
                 include_children=include_children,
                 wait_for_children=wait_for_children,
+                enabled_sections=enabled_sections,
             )
 
         return self.filter(id__in=ids)
 
 
 class EvePlanetManager(EveUniverseEntityModelManager):
-    def _fetch_from_esi(self, id) -> object:
+    def _fetch_from_esi(self, id: int, enabled_sections: Iterable[str] = None) -> dict:
         from .models import EveSolarSystem
 
-        esi_data = super()._fetch_from_esi(id)
+        esi_data = super()._fetch_from_esi(id=id)
         # no need to proceed if all children have been disabled
-        if not self.model._children():
+        if not self.model._children(enabled_sections):
             return esi_data
 
         if "system_id" not in esi_data:
             raise ValueError("system_id not found in moon response - data error")
 
         system_id = esi_data["system_id"]
-        solar_system_data = EveSolarSystem.objects._fetch_from_esi(system_id)
+        solar_system_data = EveSolarSystem.objects._fetch_from_esi(id=system_id)
         if "planets" not in solar_system_data:
             raise ValueError("planets not found in solar system response - data error")
 
@@ -474,19 +522,22 @@ class EvePlanetManager(EveUniverseEntityModelManager):
 
 
 class EvePlanetChildrenManager(EveUniverseEntityModelManager):
-    def __init__(self, property_name: str) -> None:
+    def __init__(self) -> None:
         super().__init__()
-        self._my_property_name = property_name
+        self._my_property_name = None
 
-    def _fetch_from_esi(self, id: int) -> object:
+    def _fetch_from_esi(self, id: int, enabled_sections: Iterable[str] = None) -> dict:
         from .models import EveSolarSystem
 
-        esi_data = super()._fetch_from_esi(id)
+        if not self._my_property_name:
+            raise RuntimeWarning("my_property_name not initialzed")
+
+        esi_data = super()._fetch_from_esi(id=id)
         if "system_id" not in esi_data:
             raise ValueError("system_id not found in moon response - data error")
 
         system_id = esi_data["system_id"]
-        solar_system_data = EveSolarSystem.objects._fetch_from_esi(system_id)
+        solar_system_data = EveSolarSystem.objects._fetch_from_esi(id=system_id)
         if "planets" not in solar_system_data:
             raise ValueError("planets not found in solar system response - data error")
 
@@ -505,13 +556,25 @@ class EvePlanetChildrenManager(EveUniverseEntityModelManager):
         )
 
 
+class EveAsteroidBeltManager(EvePlanetChildrenManager):
+    def __init__(self) -> None:
+        super().__init__()
+        self._my_property_name = "asteroid_belts"
+
+
+class EveMoonManager(EvePlanetChildrenManager):
+    def __init__(self) -> None:
+        super().__init__()
+        self._my_property_name = "moons"
+
+
 class EveStargateManager(EveUniverseEntityModelManager):
     """For special handling of relations"""
 
     def update_or_create_esi(
         self,
-        id: int,
         *,
+        id: int,
         include_children: bool = False,
         wait_for_children: bool = True,
         enabled_sections: Iterable[str] = None,
@@ -555,6 +618,7 @@ class EveStationManager(EveUniverseEntityModelManager):
         parent_obj: models.Model,
         inline_objects: dict,
         wait_for_children: bool,
+        enabled_sections: Iterable[str],
     ) -> None:
         """updates_or_creates station service objects for EveStations"""
         from .models import EveStationService
@@ -567,6 +631,30 @@ class EveStationManager(EveUniverseEntityModelManager):
 
             if services:
                 parent_obj.services.add(*services)
+
+
+class EveTypeManager(EveUniverseEntityModelManager):
+    def update_or_create_esi(
+        self,
+        *,
+        id: int,
+        include_children: bool = False,
+        wait_for_children: bool = True,
+        enabled_sections: Iterable[str] = None,
+    ) -> Tuple[models.Model, bool]:
+        obj, created = super().update_or_create_esi(
+            id=id,
+            include_children=include_children,
+            wait_for_children=wait_for_children,
+            enabled_sections=enabled_sections,
+        )
+        enabled_sections = self.model._enabled_sections_union(enabled_sections)
+        if enabled_sections and self.model.Section.TYPE_MATERIALS in enabled_sections:
+            from .models import EveTypeMaterial
+
+            EveTypeMaterial.objects.update_or_create_api(eve_type=obj)
+
+        return obj, created
 
 
 class EveEntityQuerySet(models.QuerySet):
@@ -826,3 +914,45 @@ class EveMarketPriceManager(models.Manager):
                 market_prices, batch_size=EVEUNIVERSE_BULK_METHODS_BATCH_SIZE
             )
             return len(market_prices)
+
+
+class EveTypeMaterialManager(models.Manager):
+    SDE_CACHE_KEY = "EVEUNIVERSE_TYPE_MATERIALS_REQUEST"
+    SDE_CACHE_TIMEOUT = 3600 * 24
+    SDE_ZZEVE_ROUTE = "invTypeMaterials.json"
+
+    def update_or_create_api(self, *, eve_type) -> None:
+        """updates or creates type material objects for the given eve type"""
+        from .models import EveType
+
+        type_material_data_all = self._fetch_sde_data_cached()
+        for type_material_data in type_material_data_all.get(eve_type.id, []):
+            material_eve_type, _ = EveType.objects.get_or_create_esi(
+                id=type_material_data.get("materialTypeID")
+            )
+            self.update_or_create(
+                eve_type=eve_type,
+                material_eve_type=material_eve_type,
+                defaults={
+                    "quantity": type_material_data.get("quantity"),
+                },
+            )
+
+    @classmethod
+    def _fetch_sde_data_cached(cls) -> dict:
+        type_material_data_all = cache.get(cls.SDE_CACHE_KEY)
+        if not type_material_data_all:
+            r = requests.get(urljoin(SDE_ZZEVE_URL, cls.SDE_ZZEVE_ROUTE))
+            r.raise_for_status()
+            type_material_data_all = dict()
+            for row in r.json():
+                type_id = row["typeID"]
+                if type_id not in type_material_data_all:
+                    type_material_data_all[type_id] = list()
+                type_material_data_all[type_id].append(row)
+            cache.set(
+                key=cls.SDE_CACHE_KEY,
+                value=type_material_data_all,
+                timeout=cls.SDE_CACHE_TIMEOUT,
+            )
+        return type_material_data_all